Added a scrollbar to the terminal windows. This is most useful for the lisp terminal window. Fixed a small bug in the scrollbar code itself. It was possible that the drag button was reduced to size 0.
Just discovered that we still have HUGE problems with the mouse cursor and flickering. 
The limitation I once implemented to prevent problems with clipping is also a very bad one.
Who fixes this once and for all? Remember that there must be a solution that works for all video drivers (DX,
OpenGL...)

- Patrick



git-svn-id: http://192.168.1.80/svn/golgotha/worksrc@51 dfb5a597-25cf-0310-bb4d-ec683fa0f420