Added dynamic language support. Whenever an scm file loaded it is now first checked wheter there exists a localized version. Localized versions have an underscore and a language identifier as suffix before the extension, ie menu_en.scm. Localized resource files (res) can be loaded by #includelang instead of just #include. But don't add the extension, it will be added automatically!
The language is choosed the following way: 
If there exists a user setting (registry or ini file) called "language" which is non-empty,
that string will be used as suffix.
Otherwise, if running windows, the system language will be used. (or better, the language of the resources in use.)
As final fallback, english is assumed.

Hint: The windows options dialog is not affected by user language settings. It always uses the system language.




git-svn-id: http://192.168.1.80/svn/golgotha/worksrc@73 dfb5a597-25cf-0310-bb4d-ec683fa0f420